src/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
distlib==0.3.1
Django==3.1.4
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
filelock==3.0.12
mysqlclient==2.0.1
Pillow==8.0.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
virtualenv==20.0.25
